Seek frequency for FFT
Now, it seeks frequency for FFT.
Not 2^n.
Numpy and Cupy can run 7^n FFT.
And so, it seeks up to 7^n wave length.
By seeking such big number, it can find the length faster.
And now, it uses cython for seeking the number.
Because, the algorithm of seeking is complicated
and could not use cuda.

PEP8 was applied.